requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.34.0
gradio>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
huggingface-hub>=0.17.0
supervision==0.18.0
openai==1.3.5
ultralytics==8.3.70
azure-identity
opencv-python-headless
dill
accelerate
timm
einops==0.8.0
paddlepaddle
paddleocr
ruff==0.6.7
pre-commit==3.8.0
pytest==8.3.3
pytest-asyncio==0.23.6
pyautogui==0.9.54
streamlit>=1.38.0
anthropic[bedrock,vertex]>=0.37.1
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
screeninfo
uiautomation
dashscope
groq